refactor: remove unused CRA template code
