+++ src/common/store/index.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import todoReducer from './todoSlice';
 
 export default configureStore({
   reducer: {
-    counter: counterReducer
+    todo: todoReducer
   }
 });
