+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
 import App from './App';
-import store from './app/store';
+import store from './common/store';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
